Moved to main and docs for GitHub pages
so in settings on GitHub it seems I don't need a special gh-pages repo
and can use plain old Main. Also as I was going to put smaller toy
projects directly in here, I thought it better to move the jekyll blog
out to /docs which is also a GitHub setting I had not read in my initail
look at how to set it up.